locus polygon algorithm with some errors

+++ app/dto/dim2D/input/InputPolygon.scala
@@ -2,7 +2,7 @@ package dto.dim2D.input
 
 import play.api.libs.json.{Json, Reads}
 
-case class InputPolygon(label: String, numberOfVertex: Int, percentage: Int, radius: Float)
+case class InputPolygon(label: String, numberOfVertex: Int, percentage: Int, radius: Double)
 
 object InputPolygon {
